+++ release/index.mjs
@@ -144,11 +144,10 @@ class InnetJS {
                 plugins: [
                     commonjs(),
                     json(),
-                    typescript2({
-                        tsconfigOverride: {
-                            compilerOptions: {
-                                declaration: false,
-                            },
+                    typescript({
+                        noEmitOnError: true,
+                        compilerOptions: {
+                            declaration: false,
                         },
                     }),
                     jsx(),
@@ -247,6 +246,7 @@ class InnetJS {
                     commonjs(),
                     json(),
                     typescript({
+                        noEmitOnError: true,
                         compilerOptions: {
                             declaration: false,
                             sourceMap: true,
@@ -314,7 +314,7 @@ class InnetJS {
                         logger.end('Bundling', e.error.message);
                         console.log(`ERROR in ${file}:${line + 1}:${column + 1}`);
                     }
-                    else if (e.error.code === 'PLUGIN_ERROR' && ['rpt2', 'commonjs'].includes(e.error.plugin)) {
+                    else if (e.error.code === 'PLUGIN_ERROR' && ['rpt2', 'commonjs', 'typescript'].includes(e.error.plugin)) {
                         const [, file, line, column] = e.error.message
                             .replace(REG_CLEAR_TEXT, '')
                             .match(REG_RPT_ERROR_FILE) || [];